chore: support get&post requests in projects api endpoint
