+++ apps/platform/components/members/MembersTableContainer.tsx
@@ -184,7 +184,7 @@ const MembersTableContainer = ({
   return (
     <div className="border-dark mt-12 w-full rounded-md border-2 shadow-sm">
       <div className="border-dark flex items-center justify-between border-b py-2 px-2 font-medium">
-        <div className="flex flex-col sm:flex-row w-full items-start gap-2 sm:items-center justify-between px-4">
+        <div className="flex w-full flex-col items-start justify-between gap-2 px-4 sm:flex-row sm:items-center">
           <div className="group relative w-full">
             <div className="text-light group-focus-within:text-lighter pointer-events-none absolute inset-y-0 left-0 flex items-center justify-center pl-3">
               <Search className="h-4 w-4" />
@@ -192,7 +192,7 @@ const MembersTableContainer = ({
             <BaseInput
               placeholder="Search members..."
               type="text"
-              className="w-full sm:max-w-xs  py-1.5 !pl-9"
+              className="w-full py-1.5  !pl-9 sm:max-w-xs"
               onChange={(e) => setQuery(e.target.value)}
               full={false}
             />
